requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make requirements.txt
#
cachetools==4.1.1         # via -r requirements.in, google-auth
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via flask, rq
deprecated==1.2.10        # via pygithub
dnspython==2.0.0          # via email-validator
email-validator==1.1.1    # via -r requirements.in
flask-caching==1.9.0      # via -r requirements.in
flask==1.1.2              # via -r requirements.in, flask-caching
gitdb==4.0.5              # via gitpython
gitpython==3.1.7          # via -r requirements.in
google-api-core==1.22.0   # via google-api-python-client
google-api-python-client==1.10.0  # via -r requirements.in
google-auth-httplib2==0.0.4  # via google-api-python-client
google-auth==1.19.2       # via -r requirements.in, google-api-core, google-api-python-client, google-auth-httplib2
googleapis-common-protos==1.52.0  # via google-api-core
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2, oauth2client
idna==2.10                # via email-validator, requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
oauth2client==4.1.3       # via -r requirements.in
protobuf==3.12.4          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth, oauth2client
pyasn1==0.4.8             # via oauth2client, pyasn1-modules, rsa
pydantic==1.6.1           # via -r requirements.in
pygithub==1.51            # via -r requirements.in
pyjwt==1.7.1              # via pygithub
python-dotenv==0.13.0     # via -r requirements.in
pytz==2020.1              # via google-api-core
pyyaml==5.3.1             # via -r requirements.in
redis==3.5.3              # via rq
requests==2.24.0          # via -r requirements.in, google-api-core, pygithub
rq==1.5.0                 # via -r requirements.in
rsa==4.6                  # via google-auth, oauth2client
six==1.15.0               # via google-api-core, google-api-python-client, google-auth, google-auth-httplib2, oauth2client, protobuf
smmap==3.0.4              # via gitdb
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.25.10          # via requests
werkzeug==1.0.1           # via flask
wrapt==1.12.1             # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
